extract delegate - JsonPrettyPrinter
