+++ src/test/java/com/ktown4u/utils/testprinter/PrettyJsonJsonPrettyPrinterTest.java
@@ -1,54 +1,23 @@
 package com.ktown4u.utils.testprinter;
 
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.PropertyAccessor;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateDeserializer;
-import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
-import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateSerializer;
-import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
 import org.approvaltests.Approvals;
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 import static com.ktown4u.utils.testprinter.OrderLineItemBuilder.anOrderLineItem;
 import static com.ktown4u.utils.testprinter.ProductBuilder.aProduct;
 import static java.util.stream.Collectors.joining;
 
-public class PrettyJsonPrinterTest {
-    private ObjectMapper mapper;
+public class PrettyJsonJsonPrettyPrinterTest {
+    private static JsonPrettyPrinter printer;
 
-    @BeforeEach
-    void setUp() {
-        createMapper();
-    }
-
-    private ObjectMapper createMapper() {
-        mapper = new ObjectMapper();
-        JavaTimeModule javaTimeModule = new JavaTimeModule();
-        javaTimeModule.addSerializer(LocalDate.class, new LocalDateSerializer(DateTimeFormatter.ofPattern("yyyy-MM-dd")));
-        javaTimeModule.addDeserializer(LocalDate.class, new LocalDateDeserializer(DateTimeFormatter.ofPattern("yyyy-MM-dd")));
-        javaTimeModule.addSerializer(LocalDateTime.class, new LocalDateTimeSerializer(DateTimeFormatter.ofPattern("yyyy-MM-dd")));
-        javaTimeModule.addDeserializer(LocalDateTime.class, new LocalDateTimeDeserializer(DateTimeFormatter.ofPattern("yyyy-MM-dd")));
-
-        mapper.disable(SerializationFeature.FAIL_ON_EMPTY_BEANS);
-        mapper.setVisibility(PropertyAccessor.ALL, JsonAutoDetect.Visibility.NONE); // Disable auto-detection for all methods
-        mapper.setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.ANY); // Enable visibility for fields
-
-        mapper.registerModule(javaTimeModule);
-
-        return mapper;
+    @BeforeAll
+    static void beforeAll() {
+        printer = new JsonPrettyPrinter();
     }
 
     @Test
@@ -70,21 +39,11 @@ public class PrettyJsonPrinterTest {
                 )
                 .build();
 
-        Approvals.verify(prettyPrint(order).stream()
+        Approvals.verify(printer.prettyPrint(order).stream()
                 .filter(line -> !line.contains("id"))
                 .filter(line -> !line.contains("description"))
                 .collect(joining("\n"))
         );
-    }
-
-    private List<String> prettyPrint(Object object) {
-        try {
-            return Arrays.asList(
-                    mapper.writerWithDefaultPrettyPrinter().writeValueAsString(object)
-                            .split("\n"));
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-        }
     }
 }
